split orderbook client functionality into different files
